banki/face-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
face-recognition==1.3.0
Pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.8.2
